scripts/pyproject.toml
[project]
name = "remove-bg"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pillow", "numpy"]
